refactor: cleanse and start from scratch
