--- src/main/kotlin/com/alpascal/api/shoppingapi/records/VideoSearchResults.kt
@@ -1,11 +1,0 @@
-package com.alpascal.api.shoppingapi.records
-
-data class VideoSearchResults(
-    val videoSearchListings: List<Video>,
-) {
-    companion object {
-        @JvmStatic
-        fun videoSearch(): List<Video> = Video.getAll()
-    }
-}
-
