--- src/test/kotlin/com/alpascal/api/shoppingapi/ShoppingApiApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.alpascal.api.shoppingapi
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class ShoppingApiApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
